% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freundlichanalysis.R
\name{freundlichanalysis}
\alias{freundlichanalysis}
\title{Describes the conformity of the results from a static adsorption experiment
to the Freundlich Isotherm model}
\usage{
freundlichanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{A numeric vector consisting of equilibrium concentration values}

\item{Qe}{A numeric vector consisting of quantities adsorbed at equilibrium}
}
\value{
The residuals, y - intercept, coefficient, and correlation factor of
the LSRL based on the Temkin isotherm model
}
\description{
Describes the conformity of the results from a static adsorption experiment
to the Freundlich Isotherm model
}
\examples{
freundlichanalysis(c(1,2,3,4), c(1,2,3,4))
}
