% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langmuirplot.R
\name{langmuirplot}
\alias{langmuirplot}
\title{Plots a Langmuir Isotherm using results from a static adsorption experiment}
\usage{
langmuirplot(Ce, Qe)
}
\arguments{
\item{Ce}{A numeric vector consisting of equilibrium concentration values}

\item{Qe}{A numeric vector consisting of quantities adsorbed at equilibrium}
}
\value{
The plot of 1 over Ce against 1 over Qe with the LSRL displayed
}
\description{
Plots a Langmuir Isotherm using results from a static adsorption experiment
}
\examples{
langmuirplot(c(1,2,3,4), c(1,2,3,4))
}
