% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryplots.R
\name{summaryplots}
\alias{summaryplots}
\title{Plots the three different Isotherm models using results from a static
adsorption experiment}
\usage{
summaryplots(Ce, Qe)
}
\arguments{
\item{Ce}{A numeric vector consisting of equilibrium concentration values}

\item{Qe}{A numeric vector consisting of quantities adsorbed at equilibrium}
}
\value{
The plots of Langmuir, Freundlich, and Temkin Isotherms
}
\description{
Plots the three different Isotherm models using results from a static
adsorption experiment
}
\examples{
summaryplots(c(1,2,3,4), c(1,2,3,4))
}
