% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temkinanalysis.R
\name{temkinanalysis}
\alias{temkinanalysis}
\title{Describes the conformity of the results from a static adsorption experiment
to the Temkin Isotherm model}
\usage{
temkinanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{A numeric vector consisting of equilibrium concentration values}

\item{Qe}{A numeric vector consisting of quantities adsorbed at equilibrium}
}
\value{
The residuals, y - intercept, coefficient, and correlation factor of
the LSRL based on the Temkin isotherm model
}
\description{
Describes the conformity of the results from a static adsorption experiment
to the Temkin Isotherm model
}
\examples{
temkinanalysis(c(1,2,3,4), c(1,2,3,4))
}
