% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temkinplot.R
\name{temkinplot}
\alias{temkinplot}
\title{Plots a Temkin Isotherm using results from a static adsorption experiment}
\usage{
temkinplot(Ce, Qe)
}
\arguments{
\item{Ce}{A numeric vector consisting of equilibrium concentration values}

\item{Qe}{A numeric vector consisting of quantities adsorbed at equilibrium}
}
\value{
The plot of ln(Ce) against Qe with the LSRL displayed
}
\description{
Plots a Temkin Isotherm using results from a static adsorption experiment
}
\examples{
temkinplot(c(1,2,3,4), c(1,2,3,4))
}
